requirements.txt
# lanchain core
langchain
langchain-core

# openAI integration
langchain-openAI
openAI

# Anthropic Integration
langchain-Anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging face Integration
langchain-Huggingface
transformers
huggingface-hub

# Environment Variable Managment
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn